requirements.txt
python==3.7.4
Django==2.2.3
djangorestframework==3.10.3

